third_party/blink/web_tests/external/wpt/fetch/private-network-access/worker-fetch.https.window-expected.txt

This is a testharness.js-based test.
PASS local to local: success.
PASS private to local: failed preflight.
PASS private to local: success.
PASS private to private: success.
PASS public to local: failed preflight.
PASS public to local: success.
PASS public to private: failed preflight.
PASS public to private: success.
PASS public to public: success.
FAIL treat-as-public to local: failed preflight. assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
PASS treat-as-public to local: success.
FAIL treat-as-public to private: failed preflight. assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
FAIL treat-as-public to private: success. assert_equals: fetch error expected (undefined) undefined but got (string) "TypeError"
PASS treat-as-public to public: success.
Harness: the test ran to completion.
